+++ etcd-recipes/src/main/kotlin/io/etcd/recipes/common/ResilienceConfig.kt
@@ -24,12 +24,18 @@ package io.etcd.recipes.common
 data class ResilienceConfig(
   val watch: WatchResilience = WatchResilience.DEFAULT,
   val lease: LeaseResilience = LeaseResilience.DEFAULT,
+  val rpc: RpcResilience = RpcResilience.DEFAULT,
 ) {
   companion object {
     @JvmField
     val DEFAULT = ResilienceConfig()
 
     @JvmField
-    val DISABLED = ResilienceConfig(watch = WatchResilience.DISABLED, lease = LeaseResilience.DISABLED)
+    val DISABLED =
+      ResilienceConfig(
+        watch = WatchResilience.DISABLED,
+        lease = LeaseResilience.DISABLED,
+        rpc = RpcResilience.DISABLED,
+      )
   }
 }